agora o cadastro de Tipo de Rating também tem limitação de permissão

+++ app/Http/Controllers/TipoRatingController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\TipoRating;
 use App\TipoRatingRegras;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\MessageBag;
 
 class TipoRatingController extends Controller
@@ -15,14 +16,29 @@ class TipoRatingController extends Controller
     }
     public function index()
     {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         $tipos_rating = TipoRating::all();
         return view("tiporating.index", compact("tipos_rating"));
     }
     function new () {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         return view('tiporating.new');
     }
     public function new_post(Request $request)
     {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         $tipo_rating = new TipoRating;
         $tipo_rating->name = $request->input("name");
         $tipo_rating->save();
@@ -30,11 +46,21 @@ class TipoRatingController extends Controller
     }
     public function dashboard($id)
     {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         $tipo_rating = TipoRating::find($id);
         return view('tiporating.dashboard', compact("tipo_rating"));
     }
     public function dashboard_post($id, Request $request)
     {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         $tipo_rating = TipoRating::find($id);
         $tipo_rating->name = $request->input("name");
         $tipo_rating->save();
@@ -42,6 +68,11 @@ class TipoRatingController extends Controller
     }
     public function delete($id)
     {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         $tipo_rating = TipoRating::find($id);
 
         if ($tipo_rating->isDeletavel()) {
@@ -52,6 +83,11 @@ class TipoRatingController extends Controller
 
     public function regra_add($id, Request $request)
     {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         if (!$request->has("idade_minima") && !$request->has("idade_maxima")) {
             $messageBag = new MessageBag;
             $messageBag->add("alerta", "É necessário definir ao menos uma idade mínima ou máxima, ou então as duas, para a regra em questão.");
@@ -85,6 +121,11 @@ class TipoRatingController extends Controller
     }
     public function regra_remove($id, $tipo_rating_regra_id)
     {
+        $user = Auth::user();
+        if (!$user->hasPermissionGlobalbyPerfil([1, 2])) {
+            return redirect("/");
+        }
+
         $tipo_rating_regra = TipoRatingRegras::find($tipo_rating_regra_id);
         $tipo_rating_regra->delete();
         return redirect("/tiporating/dashboard/" . $id);
